requirements.txt
annotated-types==0.7.0
pandas==2.2.3
python-jobspy==1.1.80
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
Jinja2==3.0.3
aiohttp==3.9.5